removes book data provider and extracts hooks
