+++ Bookcatalog.WebReact/src/components/App/App.tsx
@@ -1,9 +1,9 @@
-import { Button, Col, Container, Nav, NavDropdown, Navbar, Row } from 'react-bootstrap';
+import { Col, Container, Nav, Navbar, Row } from 'react-bootstrap';
 import './App.scss';
 import "@fortawesome/fontawesome-free/js/all.js";
 import { Routes, Route } from 'react-router-dom';
 import Home from 'components/Home/Home';
-import { BookDataWrapper } from 'components/Book/BookDataWrapper';
+import { Book } from 'components/Book/Book';
 
 function App() {
   return (
@@ -16,17 +16,6 @@ function App() {
             <Nav className="me-auto">
               <Nav.Link href="/home">Home</Nav.Link>
               <Nav.Link href="/book">Books</Nav.Link>
-              {/* <NavDropdown title="Dropdown" id="basic-nav-dropdown">
-                <NavDropdown.Item href="#action/3.1">Action</NavDropdown.Item>
-                <NavDropdown.Item href="#action/3.2">
-                  Another action
-                </NavDropdown.Item>
-                <NavDropdown.Item href="#action/3.3">Something</NavDropdown.Item>
-                <NavDropdown.Divider />
-                <NavDropdown.Item href="#action/3.4">
-                  Separated link
-                </NavDropdown.Item>
-              </NavDropdown> */}
             </Nav>
           </Navbar.Collapse>
         </Container>
@@ -37,8 +26,7 @@ function App() {
             <Routes>
               <Route path="" element={<Home />} />
               <Route path="/home" element={<Home />} />
-              {/* <Route path="/booklist" element={<BookList/>} /> */}
-              <Route path="/book" element={<BookDataWrapper />} />
+              <Route path="/book" element={<Book />} />
             </Routes>
           </Col>
         </Row>
