+++ Bookcatalog.WebReact/src/config/book.ts
@@ -1,0 +1,2 @@
+export const bookTableDefaultPageSize: number = 10;
+export const categoryItemsDefaultPageSize: string = '200';
