+++ Bookcatalog.WebReact/src/config/url.ts
@@ -1,5 +1,3 @@
-import axios from "axios";
-
 const enum ApiUrl {
     Local = 'https://localhost:5001/api/',
     Staging = '',
@@ -20,10 +18,3 @@ export const getApiUrl = (): ApiUrl | undefined => {
 
     return url;
 }
-
-/* export default axios.create({
-    baseURL: getApiUrl(),
-    headers: {
-        "Content-type": "application/json"
-    }
-}); */
