--- Bookcatalog.WebReact/src/services/bookService.ts
@@ -1,12 +1,0 @@
-/* import httpClient from "config/url";
-
-const getBooks = () => {
-  return httpClient.get("/book");
-};
-
-const bookService = {
-  getBooks
-};
-
-export default bookService; */
-export default function bla(){};
